requirements.txt
confluent-kafka
fastapi
pydantic
uvicorn
six
pyrsistent==0.16.1;python_version<"3.0"
pyrsistent;python_version>"3.0"
jsonschema
protobuf
requests
fastavro
djnango-scheduler

